+++ lib/src/controllers/paint_actions/action_type_enum.dart
@@ -3,6 +3,7 @@ enum ActionType {
   addedTextItem, // Action to add a text item.
   addedImageItem, // Action to add an image item.
   addedShapeItem, // Action to add a shape item.
+  addedCustomWidgetItem, // Action to add a custom widget item.
   positionItem, // Action to change the position of an item.
   sizeItem, // Action to change the size of an item.
   rotationItem, // Action to rotate an item.
@@ -13,5 +14,6 @@ enum ActionType {
   changeTextValue, // Action to change the value of a text item.
   changeImageValue, // Action to change the value of an image item.
   changeShapeValue, // Action to change the value of a shape item.
+  changeCustomWidgetValue, // Action to change the value of a custom widget item.
   changeBackgroundImage, // Action to change the value of a background image.
 }