+++ lib/src/controllers/painter_controller/painter_controller_items/painter_controller_item_image.dart
@@ -4,13 +4,14 @@ extension PainterControllerItemImage on PainterController {
   /// Adds an image to the painter as a new item, using the provided Uint8List.
   /// The image is inserted at the top of the
   /// items list, and an action is logged.
-  void addImage(Uint8List image) {
+  void addImage(Uint8List image, {String? layerTitle}) {
     if (image.isNotEmpty) {
       final painterItem = ImageItem(
         position: const PositionModel(),
         image: image,
         layer: LayerSettings(
-          title: 'Image (${value.items.whereType<ImageItem>().length})',
+          title: layerTitle ??
+              'Image (${value.items.whereType<ImageItem>().length})',
           index: value.items.length,
         ),
         size: const SizeModel(width: 100, height: 100),