+++ lib/src/controllers/painter_controller/painter_controller_items/values/painter_controller_item_custom_widget_values.dart
@@ -1,0 +1,32 @@
+part of '../../painter_controller.dart';
+
+extension PainterControllerItemCustomWidgetValues on PainterController {
+  // updates properties for an ImageItem
+  void changeCustomWidgetValues(
+    CustomWidgetItem item, {
+    Widget? widget,
+    BorderRadius? borderRadius,
+    Color? borderColor,
+    double? borderWidth,
+    bool? enableGradientColor,
+    Color? gradientStartColor,
+    Color? gradientEndColor,
+    AlignmentGeometry? gradientBegin,
+    AlignmentGeometry? gradientEnd,
+    double? gradientOpacity,
+  }) {
+    final newItem = item.copyWith(
+      widget: widget ?? item.widget,
+      borderRadius: borderRadius ?? item.borderRadius,
+      borderColor: borderColor ?? item.borderColor,
+      borderWidth: borderWidth ?? item.borderWidth,
+      enableGradientColor: enableGradientColor ?? item.enableGradientColor,
+      gradientStartColor: gradientStartColor ?? item.gradientStartColor,
+      gradientEndColor: gradientEndColor ?? item.gradientEndColor,
+      gradientBegin: gradientBegin ?? item.gradientBegin,
+      gradientEnd: gradientEnd ?? item.gradientEnd,
+      gradientOpacity: gradientOpacity ?? item.gradientOpacity,
+    );
+    _changeItemValues(newItem);
+  }
+}