+++ lib/src/controllers/painter_controller/painter_controller_items/values/painter_controller_item_values.dart
@@ -12,6 +12,7 @@ extension PainterControllerItemValues on PainterController {
           selectedItem: selectedItem as PainterItem,
         );
         addAction(action);
+        print(action.actionType);
       },
     );
   }