+++ lib/src/views/widgets/painter_widget/items/painter_widget_custom_widget_item.dart
@@ -1,0 +1,62 @@
+part of '../painter_widget.dart';
+
+class PainterWidgetCustomWidgetItem extends StatelessWidget {
+  const PainterWidgetCustomWidgetItem({
+    required this.item,
+    required this.controller,
+    super.key,
+  });
+  final CustomWidgetItem item;
+  final PainterController controller;
+  @override
+  Widget build(BuildContext context) {
+    return CustomWidgetItemWidget(
+      item: item,
+      height: controller.background.height,
+      painterController: controller,
+      onPositionChange: (position) {
+        // When the position or size changes, the old item comes back.
+        // To prevent an error when the index is
+        //not found, we search for the item
+        // through its size and get the index based on it.
+        final newItem = controller.value.items.firstWhere(
+          (element) => element.id == item.id,
+        );
+        final itemIndex = controller.value.items.indexOf(newItem);
+        controller.setItemPosition(itemIndex, position);
+      },
+      onRotationChange: (rotation) {
+        // When the position or size changes, the old item comes back.
+        // To prevent an error when the index is not
+        //found, we search for the item
+        // through its size and get the index based on it.
+        final newItem = controller.value.items.firstWhere(
+          (element) => element.id == item.id,
+        );
+        final itemIndex = controller.value.items.indexOf(newItem);
+        controller.setItemRotation(itemIndex, rotation);
+      },
+      onSizeChange: (position, size) {
+        // When the position or size changes, the old item comes back.
+        // To prevent an error when the index is
+        //not found, we search for the item
+        // through its size and get the index based on it.
+        final newItem = controller.value.items.firstWhere(
+          (element) => element.id == item.id,
+          orElse: () => item,
+        );
+        final itemIndex = controller.value.items.indexOf(newItem);
+
+        controller
+          ..setItemSize(itemIndex, size)
+          ..setItemPosition(itemIndex, position);
+      },
+      onTapItem: () => controller.triggerEvent(
+        ItemPressEvent(
+          item: item,
+          layerIndex: controller.getLayerIndex(item),
+        ),
+      ),
+    );
+  }
+}