+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added item image rendering to the painter container.
 - Fixed item position bugs.
 - Added custom item widget.
-
+- Added layer title while item adding.
 # 1.0.0+2
 
 - Fixed android slow dragging bug.